Implement derivers for decoders
