+++ modules/core/src/main/scala/me/mnedokushev/zio/apache/arrow/core/codec/ValueEncoder.scala
@@ -1,16 +1,15 @@
 package me.mnedokushev.zio.apache.arrow.core.codec
 
-import org.apache.arrow.vector.ValueVector
+import org.apache.arrow.memory.{ ArrowBuf, BufferAllocator }
 import org.apache.arrow.vector.complex.writer.FieldWriter
+import org.apache.arrow.vector.{ ValueVector, _ }
 import zio.Chunk
-import org.apache.arrow.memory.BufferAllocator
-import zio.schema.StandardType
-import org.apache.arrow.vector._
+import zio.schema.{ Schema, StandardType }
+
 import java.nio.ByteBuffer
-import java.util.UUID
-import java.time._
 import java.nio.charset.StandardCharsets
-import zio.schema.Schema
+import java.time._
+import java.util.UUID
 
 trait ValueEncoder[-A] {
 
@@ -140,5 +139,111 @@ object ValueEncoder {
       case (other, _, _)                                                                =>
         throw EncoderError(s"Unsupported ZIO Schema StandardType $other")
     }
+
+  private[codec] def encodePrimitive[A](
+    standardType: StandardType[A],
+    value: A,
+    name: Option[String],
+    writer: FieldWriter
+  )(implicit alloc: BufferAllocator): Unit = {
+
+    def withBuffer(size: Long)(fn: ArrowBuf => Unit) = {
+      val buffer = alloc.buffer(size)
+      fn(buffer)
+      buffer.close()
+    }
+
+    def writeString(s: String) =
+      withBuffer(s.length.toLong) { buffer =>
+        buffer.writeBytes(s.getBytes(StandardCharsets.UTF_8))
+        name.fold(writer.varChar)(writer.varChar).writeVarChar(0, s.length, buffer)
+      }
+
+    def writeLong[A1](v: A1)(fst: A1 => Int)(snd: A1 => Int) =
+      withBuffer(8) { buffer =>
+        buffer.writeInt(fst(v))
+        buffer.writeInt(snd(v))
+        name.fold(writer.bigInt)(writer.bigInt).writeBigInt(buffer.getLong(0))
+      }
+
+    (standardType, value) match {
+      case (StandardType.StringType, v: String)                   =>
+        writeString(v)
+      case (StandardType.BoolType, v: Boolean)                    =>
+        name.fold(writer.bit)(writer.bit).writeBit(if (v) 1 else 0)
+      case (StandardType.ByteType, v: Byte)                       =>
+        name.fold(writer.uInt1)(writer.uInt1).writeUInt1(v)
+      case (StandardType.ShortType, v: Short)                     =>
+        name.fold(writer.smallInt)(writer.smallInt).writeSmallInt(v)
+      case (StandardType.IntType, v: Int)                         =>
+        name.fold(writer.integer)(writer.integer).writeInt(v)
+      case (StandardType.LongType, v: Long)                       =>
+        name.fold(writer.bigInt)(writer.bigInt).writeBigInt(v)
+      case (StandardType.FloatType, v: Float)                     =>
+        name.fold(writer.float4)(writer.float4).writeFloat4(v)
+      case (StandardType.DoubleType, v: Double)                   =>
+        name.fold(writer.float8)(writer.float8).writeFloat8(v)
+      case (StandardType.BinaryType, v: Chunk[_])                 =>
+        withBuffer(v.length.toLong) { buffer =>
+          buffer.writeBytes(v.asInstanceOf[Chunk[Byte]].toArray)
+          name.fold(writer.largeVarBinary)(writer.largeVarBinary).writeLargeVarBinary(0L, v.length.toLong, buffer)
+        }
+      case (StandardType.CharType, v: Char)                       =>
+        name.fold(writer.uInt2)(writer.uInt2).writeUInt2(v)
+      case (StandardType.UUIDType, v: UUID)                       =>
+        withBuffer(16) { buffer =>
+          buffer.writeLong(v.getLeastSignificantBits)
+          buffer.writeLong(v.getMostSignificantBits)
+          name.fold(writer.varBinary)(writer.varBinary).writeVarBinary(0, 16, buffer)
+        }
+      case (StandardType.BigDecimalType, v: java.math.BigDecimal) =>
+        name.fold(writer.decimal)(writer.decimal).writeDecimal(v)
+      case (StandardType.BigIntegerType, v: java.math.BigInteger) =>
+        val bb = v.toByteArray
+        withBuffer(bb.length.toLong) { buffer =>
+          buffer.writeBytes(bb)
+          name.fold(writer.varBinary)(writer.varBinary).writeVarBinary(0, bb.length, buffer)
+        }
+      case (StandardType.DayOfWeekType, v: DayOfWeek)             =>
+        name.fold(writer.integer)(writer.integer).writeInt(v.getValue)
+      case (StandardType.MonthType, v: Month)                     =>
+        name.fold(writer.integer)(writer.integer).writeInt(v.getValue)
+      case (StandardType.MonthDayType, v: MonthDay)               =>
+        writeLong(v)(_.getDayOfMonth)(_.getMonthValue)
+      case (StandardType.PeriodType, v: Period)                   =>
+        withBuffer(12) { buffer =>
+          buffer.writeInt(v.getDays)
+          buffer.writeInt(v.getMonths)
+          buffer.writeInt(v.getYears)
+          name.fold(writer.varBinary)(writer.varBinary).writeVarBinary(0, 12, buffer)
+        }
+      case (StandardType.YearType, v: Year)                       =>
+        name.fold(writer.integer)(writer.integer).writeInt(v.getValue)
+      case (StandardType.YearMonthType, v: YearMonth)             =>
+        writeLong(v)(_.getMonthValue)(_.getYear)
+      case (StandardType.ZoneIdType, v: ZoneId)                   =>
+        writeString(v.toString)
+      case (StandardType.ZoneOffsetType, v: ZoneOffset)           =>
+        writeString(v.toString)
+      case (StandardType.DurationType, v: Duration)               =>
+        name.fold(writer.bigInt)(writer.bigInt).writeBigInt(v.toMillis)
+      case (StandardType.InstantType, v: Instant)                 =>
+        name.fold(writer.bigInt)(writer.bigInt).writeBigInt(v.toEpochMilli)
+      case (StandardType.LocalDateType, v: LocalDate)             =>
+        writeString(v.toString)
+      case (StandardType.LocalTimeType, v: LocalTime)             =>
+        writeString(v.toString)
+      case (StandardType.LocalDateTimeType, v: LocalDateTime)     =>
+        writeString(v.toString)
+      case (StandardType.OffsetTimeType, v: OffsetTime)           =>
+        writeString(v.toString)
+      case (StandardType.OffsetDateTimeType, v: OffsetDateTime)   =>
+        writeString(v.toString)
+      case (StandardType.ZonedDateTimeType, v: ZonedDateTime)     =>
+        writeString(v.toString)
+      case (other, _)                                             =>
+        throw EncoderError(s"Unsupported ZIO Schema StandardType $other")
+    }
+  }
 
 }
