+++ modules/core/src/main/scala/me/mnedokushev/zio/apache/arrow/core/codec/VectorSchemaRootEncoderDeriver.scala
@@ -1,15 +1,12 @@
 package me.mnedokushev.zio.apache.arrow.core.codec
 
-import zio.schema.Deriver
-import zio.schema.Schema
-import zio.Chunk
-import org.apache.arrow.vector.VectorSchemaRoot
 import org.apache.arrow.memory.BufferAllocator
+import org.apache.arrow.vector.complex.impl.{ PromotableWriter, UnionListWriter }
 import org.apache.arrow.vector.complex.writer.FieldWriter
-import org.apache.arrow.vector.complex.StructVector
-import org.apache.arrow.vector.complex.ListVector
-import org.apache.arrow.vector.FieldVector
-import zio.schema.StandardType
+import org.apache.arrow.vector.complex.{ ListVector, StructVector }
+import org.apache.arrow.vector.{ FieldVector, VectorSchemaRoot }
+import zio.Chunk
+import zio.schema.{ Deriver, Schema, StandardType }
 
 object VectorSchemaRootEncoderDeriver {
 
@@ -21,14 +18,10 @@ object VectorSchemaRootEncoderDeriver {
       summoned: => Option[VectorSchemaRootEncoder[A]]
     ): VectorSchemaRootEncoder[A] = new VectorSchemaRootEncoder[A] {
 
-      override def encodeField(name: String, vec: FieldVector, writer: FieldWriter, value: A, idx: Int)(implicit
-        alloc: BufferAllocator
-      ): Unit =
-        ValueEncoder.encodeStruct(value, record.fields, fields.map(_.unwrap), writer)
+      private val encoders = fields.map(_.unwrap)
 
       private def encodeField0[A1](
         encoder: VectorSchemaRootEncoder[_],
-        name: String,
         vec: FieldVector,
         writer: FieldWriter,
         value: A1,
@@ -36,14 +29,14 @@ object VectorSchemaRootEncoderDeriver {
       )(implicit
         alloc: BufferAllocator
       ) =
-        encoder.asInstanceOf[VectorSchemaRootEncoder[A1]].encodeField(name, vec, writer, value, idx)
+        encoder.asInstanceOf[VectorSchemaRootEncoder[A1]].encodeField(vec, writer, value, idx)
 
       override protected def encodeUnsafe(
         chunk: Chunk[A],
         root: VectorSchemaRoot
       )(implicit alloc: BufferAllocator): VectorSchemaRoot = {
         val fields0 =
-          record.fields.zip(fields.map(_.unwrap)).map { case (Schema.Field(name, fieldSchema, _, _, g, _), encoder) =>
+          record.fields.zip(encoders).map { case (Schema.Field(name, fieldSchema, _, _, g, _), encoder) =>
             val vec = Option(root.getVector(name))
               .getOrElse(throw EncoderError(s"Couldn't find vector by name $name"))
 
@@ -55,19 +48,19 @@ object VectorSchemaRootEncoderDeriver {
               case _                                               => null
             }
 
-            (encoder, name, vec, writer, g)
+            (encoder, vec, writer, g)
           }
 
         val len = chunk.length
         val it  = chunk.iterator.zipWithIndex
 
         it.foreach { case (v, i) =>
-          fields0.foreach { case (encoder, name, vec, writer, get) =>
-            encodeField0(encoder, name, vec, writer, get(v), i)
+          fields0.foreach { case (encoder, vec, writer, get) =>
+            encodeField0(encoder, vec, writer, get(v), i)
           }
         }
 
-        fields0.foreach { case (_, _, vec, _, _) =>
+        fields0.foreach { case (_, vec, _, _) =>
           vec.setValueCount(len)
         }
 
@@ -75,10 +68,22 @@ object VectorSchemaRootEncoderDeriver {
         root
       }
 
+      def encodeValue(
+        value: A,
+        name: Option[String],
+        writer: FieldWriter
+      )(implicit alloc: BufferAllocator): Unit = {
+        val writer0 = name.fold[FieldWriter](writer.struct().asInstanceOf[UnionListWriter])(
+          writer.struct(_).asInstanceOf[PromotableWriter]
+        )
+
+        ValueEncoder.encodeStruct(value, record.fields, encoders, writer0)
+      }
+
     }
 
     override def deriveEnum[A](
-      enum: Schema.Enum[A],
+      `enum`: Schema.Enum[A],
       cases: => Chunk[Deriver.WrappedF[VectorSchemaRootEncoder, _]],
       summoned: => Option[VectorSchemaRootEncoder[A]]
     ): VectorSchemaRootEncoder[A] = ???
@@ -88,7 +93,12 @@ object VectorSchemaRootEncoderDeriver {
       summoned: => Option[VectorSchemaRootEncoder[A]]
     ): VectorSchemaRootEncoder[A] = new VectorSchemaRootEncoder[A] {
 
-      override def encodeField(name: String, vec: FieldVector, writer: FieldWriter, value: A, idx: Int)(implicit
+      override def encodeValue(value: A, name: Option[String], writer: FieldWriter)(implicit
+        alloc: BufferAllocator
+      ): Unit =
+        ValueEncoder.encodePrimitive(st, value, name, writer)
+
+      override def encodeField(vec: FieldVector, writer: FieldWriter, value: A, idx: Int)(implicit
         alloc: BufferAllocator
       ): Unit =
         ValueEncoder.encodePrimitive(st, value, vec, idx)
@@ -107,10 +117,13 @@ object VectorSchemaRootEncoderDeriver {
       summoned: => Option[VectorSchemaRootEncoder[C[A]]]
     ): VectorSchemaRootEncoder[C[A]] = new VectorSchemaRootEncoder[C[A]] {
 
-      override def encodeField(name: String, vec: FieldVector, writer: FieldWriter, value: C[A], idx: Int)(implicit
+      override def encodeValue(value: C[A], name: Option[String], writer: FieldWriter)(implicit
         alloc: BufferAllocator
-      ): Unit =
-        ValueEncoder.encodeList(sequence.toChunk(value), inner, writer)
+      ): Unit = {
+        val writer0 = name.fold(writer.list)(writer.list).asInstanceOf[PromotableWriter]
+
+        ValueEncoder.encodeList(sequence.toChunk(value), inner, writer0)
+      }
 
     }
